SSLR-205 Matchers should control creation of nodes
Thus, the responsibility for the decision whether or not to create a
node of the parse tree is fully under control of matchers and context is
used only as a shared state. So there is no need to create nodes from
matchers for "leafs" (PatternMatcher, StringMatcher, EndOfInputMatcher)
and for "predicates" (TestMatcher, TestNotMatcher) as they never
contribute to the parse tree.